api用のapplicationを作成し、restframeworkに関わるviewとserializerモジュールを移行。ajaxはapiへのルーティングへ変更
