+++ app/account/urls.py
@@ -1,17 +1,10 @@
 from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
-
-from rest_framework.routers import DefaultRouter
-
 from . import views
 
 
 app_name = 'account'
-router = DefaultRouter()
-router.register('profileEdit', views.UserProfileEditView)
-router.register('relation', views.RelationView)
-
 
 urlpatterns = [
     path('login/', views.Login.as_view(), name="login"),
@@ -23,5 +16,4 @@ urlpatterns = [
          views.FollowingView.as_view(), name='following'),
     path('like/<int:user_id>', views.LikeArticleView.as_view(), name='like'),
     path('media/<int:user_id>', views.ArticleMedia.as_view(), name='media'),
-    path('', include(router.urls)),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
