+++ app/app/urls.py
@@ -24,4 +24,5 @@ urlpatterns = [
     path('article/', include('article.urls')),
     path('account/', include('account.urls')),
     path('messages/', include('message.urls')),
+    path('api/', include('api.urls')),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT) + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
