+++ app/templates/article/media.html
@@ -111,7 +111,7 @@
       //ユーザのプロフィールをgetするための処理
       $('#id_user_profile').on('click',function(){
         $.ajax({
-          url:'http://localhost/account/profileEdit/',
+          url:'http://localhost/api/profileEdit/',
           type:'GET',
           headers: {csrfmiddlewaretoken: 'Token ' +window.CSRF_TOKEN},
           success: afterSuccessGetProfile,
@@ -155,7 +155,7 @@
           return;
         }
         $.ajax({
-          url:'http://localhost/account/profileEdit/{{user_profile.id}}/',
+          url:'http://localhost/api/profileEdit/{{user_profile.id}}/',
           type:'PATCH',
           dataType:'json',
           data:data,
@@ -189,7 +189,7 @@
         var data = new FormData();
         data.append('target',target);
         $.ajax({
-          url:'http://localhost/account/relation/',
+          url:'http://localhost/api/relation/',
           type:'POST',
           dataType:'json',
           data:data,
@@ -201,7 +201,7 @@
       $('#id_unrelate_button').on('click',function(){
         var relationId = $("#id_relation_id_port").val()
         $.ajax({
-          url:'http://localhost/account/relation/'+relationId+'/',
+          url:'http://localhost/api/relation/'+relationId+'/',
           type:'DELETE',
           success:afterSuccessUnfollow,
           processData:false,
