+++ app/templates/article/profile.html
@@ -127,7 +127,7 @@
       //ユーザのプロフィールをgetするための処理
       $('#id_user_profile').on('click',function(){
         $.ajax({
-          url:'http://localhost/account/profileEdit/',
+          url:'http://localhost/api/profileEdit/',
           type:'GET',
           headers: {csrfmiddlewaretoken: 'Token ' +window.CSRF_TOKEN},
           success: afterSuccessGetProfile,
@@ -159,7 +159,7 @@
           data.append('profile_image',profileImage);
         }
         if(backgroundImage){
-          var backgroundExtension = background_image.name.split(".").pop()
+          var backgroundExtension = backgroundImage.name.split(".").pop()
             if(backgroundExtension!='png'&& backgroundExtension!='jpg'){
               alert("背景画像の拡張子はjpegもしくはpngでアップロードしてください")
               return 
@@ -172,7 +172,7 @@
           return;
         }
         $.ajax({
-          url:'http://localhost/account/profileEdit/{{user_profile.id}}/',
+          url:'http://localhost/api/profileEdit/{{user_profile.id}}/',
           type:'PATCH',
           dataType:'json',
           data:data,
@@ -206,7 +206,7 @@
         var data = new FormData();
         data.append('target',target);
         $.ajax({
-          url:'http://localhost/account/relation/',
+          url:'http://localhost/api/relation/',
           type:'POST',
           dataType:'json',
           data:data,
@@ -218,7 +218,7 @@
       $('#id_unrelate_button').on('click',function(){
         var relationId = $("#id_relation_id_port").val()
         $.ajax({
-          url:'http://localhost/account/relation/'+relationId+'/',
+          url:'http://localhost/api/relation/'+relationId+'/',
           type:'DELETE',
           success:afterSuccessUnfollow,
           processData:false,
